+++ actix-web/CHANGES.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Add `resources-introspection` feature for retrieving configured route paths and HTTP methods.
+
 ## 4.10.2
 
 - No significant changes since `4.10.1`.